requirements.txt
Arrays==0.0.5
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-ckeditor==6.5.1
django-cleanup==8.0.0
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.23.4
openpyxl==3.0.10
packaging==23.1
pandas==1.5.1
Pillow==9.5.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
